+++ packages/interfaces/src/__test__/IAppliance.test.js
@@ -1,3 +1,6 @@
+// We test logging in this suite
+/* eslint-disable no-console */
+
 import {
 	InterfaceInstantiationError,
 	NotImplementedError,
@@ -25,6 +28,83 @@ describe('IAppliance', () => {
 			expect(() => {
 				new PartiallyImplementedAppliance() // eslint-disable-line no-new
 			}).not.toThrow(Error)
+		})
+	})
+
+	describe('logger', () => {
+		it('should be defined', () => {
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toBeDefined()
+		})
+		it('should support the log() method', () => {
+			const consoleLogCopy = console.log
+			console.log = jest.fn()
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toHaveProperty('log')
+			expect(typeof appliance.logger.log).toBe('function')
+			expect(() => appliance.logger.log('customLevel', 'Test message')).not.toThrow()
+			expect(console.log).toHaveBeenCalledTimes(1)
+			console.log = consoleLogCopy
+		})
+		it('should support the fatal() method', () => {
+			const consoleLogCopy = console.log
+			console.log = jest.fn()
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toHaveProperty('fatal')
+			expect(typeof appliance.logger.fatal).toBe('function')
+			expect(() => appliance.logger.fatal('Test message')).not.toThrow()
+			expect(console.log).toHaveBeenCalledTimes(1)
+			console.log = consoleLogCopy
+		})
+		it('should support the error() method', () => {
+			const consoleLogCopy = console.log
+			console.log = jest.fn()
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toHaveProperty('error')
+			expect(typeof appliance.logger.error).toBe('function')
+			expect(() => appliance.logger.error('Test message')).not.toThrow()
+			expect(console.log).toHaveBeenCalledTimes(1)
+			console.log = consoleLogCopy
+		})
+		it('should support the warn() method', () => {
+			const consoleLogCopy = console.log
+			console.log = jest.fn()
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toHaveProperty('warn')
+			expect(typeof appliance.logger.warn).toBe('function')
+			expect(() => appliance.logger.warn('Test message')).not.toThrow()
+			expect(console.log).toHaveBeenCalledTimes(1)
+			console.log = consoleLogCopy
+		})
+		it('should support the info() method', () => {
+			const consoleLogCopy = console.log
+			console.log = jest.fn()
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toHaveProperty('info')
+			expect(typeof appliance.logger.info).toBe('function')
+			expect(() => appliance.logger.info('Test message')).not.toThrow()
+			expect(console.log).toHaveBeenCalledTimes(1)
+			console.log = consoleLogCopy
+		})
+		it('should support the debug() method', () => {
+			const consoleLogCopy = console.log
+			console.log = jest.fn()
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toHaveProperty('debug')
+			expect(typeof appliance.logger.debug).toBe('function')
+			expect(() => appliance.logger.debug('Test message')).not.toThrow()
+			expect(console.log).toHaveBeenCalledTimes(1)
+			console.log = consoleLogCopy
+		})
+		it('should support the trace() method', () => {
+			const consoleLogCopy = console.log
+			console.log = jest.fn()
+			const appliance = new PartiallyImplementedAppliance()
+			expect(appliance.logger).toHaveProperty('trace')
+			expect(typeof appliance.logger.trace).toBe('function')
+			expect(() => appliance.logger.trace('Test message')).not.toThrow()
+			expect(console.log).toHaveBeenCalledTimes(1)
+			console.log = consoleLogCopy
 		})
 	})